requirements.txt
requests==2.31.0
python-dotenv==1.0.0
numpy
pandas
xlrd==2.0.1
openpyxl
ynab
inquirer
openai
PyPDF2
playwright